+++ src/lib.rs
@@ -524,6 +524,7 @@ pub use codec::Codec;
 pub use de::from_value;
 pub use reader::{from_avro_datum, Reader};
 pub use schema::{ParseSchemaError, Schema};
+pub use ser::to_value;
 pub use types::SchemaResolutionError;
 pub use util::{max_allocation_bytes, DecodeError};
 pub use writer::{to_avro_datum, ValidationError, Writer};