+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## Added
+- Support for Parsing Canonical Form
+- `to_value` to serialize anything that implements Serialize into a Value
 
 ## [0.4.1] - 2018-06-17
 ### Changed